requirements.txt
alpaca-py==0.39.1
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
msgpack==1.1.0
numpy==2.2.4
pandas==2.2.3
pydantic==2.11.1
pydantic_core==2.33.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
sseclient-py==1.8.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
websockets==15.0.1
python-dotenv==1.0.0
matplotlib==3.8.3
ta==0.11.0

# Testing tools
pytest==8.0.1
pytest-cov==4.1.0
pytest-mock==3.12.0
